+++ managers/process_manager.py
@@ -29,6 +29,7 @@ from exclusiveprocess import (
 )
 import multiprocessing
 
+
 import modules
 from modules.update_manager.update_manager import UpdateManager
 from slips_files.common.slips_utils import utils
@@ -53,6 +54,9 @@ class ProcessManager:
         # to pass flows to the profiler
         self.profiler_queue = Queue()
         self.termination_event: Event = Event()
+        # to make sure we only warn the user once about
+        # the pending modules
+        self.warning_printed_once = False
         # this one has its own termination event because we want it to
         # shutdown at the very end of all other slips modules.
         self.evidence_handler_termination_event: Event = Event()
@@ -77,6 +81,9 @@ class ProcessManager:
         self.modules_to_ignore: list = self.main.conf.get_disabled_modules(
             self.main.input_type
         )
+        self.bootstrap_p2p = self.main.conf.is_bootstrapping_node()
+        self.bootstrapping_modules = self.main.conf.get_bootstrapping_modules()
+        # self.bootstrap_p2p, self.boootstrapping_modules = self.main.conf.get_bootstrapping_setting()
 
     def start_output_process(self, stderr, slips_logfile, stdout=""):
         output_process = Output(
@@ -210,6 +217,26 @@ class ProcessManager:
                 return True
         return False
 
+    def is_bootstrapping_module(self, module_name: str) -> bool:
+        m1 = (
+            module_name.replace(" ", "")
+            .replace("_", "")
+            .replace("-", "")
+            .lower()
+        )
+        for bootstrap_module in self.bootstrapping_modules:
+            m2 = (
+                bootstrap_module.replace(" ", "")
+                .replace("_", "")
+                .replace("-", "")
+                .lower()
+            )
+
+            if m1.__contains__(m2):
+                return True
+        self.modules_to_ignore.append(module_name.split(".")[-1])
+        return False
+
     def is_abstract_module(self, obj) -> bool:
         return obj.name in ("IModule", "AsyncModule")
 
@@ -241,8 +268,16 @@ class ProcessManager:
             if dir_name != file_name:
                 continue
 
-            if self.is_ignored_module(module_name):
-                continue
+            if self.bootstrap_p2p:  # if bootstrapping the p2p network
+                if not self.is_bootstrapping_module(
+                    module_name
+                ):  # keep only the bootstrapping-necessary modules
+                    continue
+            else:  # if not bootstrappig mode
+                if self.is_ignored_module(
+                    module_name
+                ):  # ignore blacklisted modules
+                    continue
 
             # Try to import the module, otherwise skip.
             try:
@@ -436,17 +471,18 @@ class ProcessManager:
             pid for pid in pids_to_kill_last if pid is not None
         ]
 
+        # now get the process obj of each pid
         to_kill_first: List[Process] = []
         to_kill_last: List[Process] = []
         for process in self.processes:
             if process.pid in pids_to_kill_last:
                 to_kill_last.append(process)
-            else:
+            elif isinstance(process, multiprocessing.context.ForkProcess):
                 # skips the context manager of output.py, will close
                 # it manually later
                 # once all processes are closed
-                if isinstance(process, multiprocessing.context.ForkProcess):
-                    continue
+                continue
+            else:
                 to_kill_first.append(process)
 
         return to_kill_first, to_kill_last
@@ -645,12 +681,14 @@ class ProcessManager:
                 print("\n" + "-" * 27)
             print("Stopping Slips")
 
-            # by default, 15 mins from this time, all modules should be killed
+            # by default, max 15 mins (taken from wait_for_modules_to_finish)
+            # from this time, all modules should be killed
             method_start_time = time.time()
 
             # how long to wait for modules to finish in minutes
             timeout: float = self.main.conf.wait_for_modules_to_finish()
-            timeout_seconds: float = timeout * 60
+            # convert to seconds
+            timeout *= 60
 
             # close all tws
             self.main.db.check_tw_to_close(close_all=True)
@@ -669,20 +707,15 @@ class ProcessManager:
                     log_to_logfiles_only=True,
                 )
 
-                hitlist: Tuple[List[Process], List[Process]]
-                hitlist = self.get_hitlist_in_order()
-                to_kill_first: List[Process] = hitlist[0]
-                to_kill_last: List[Process] = hitlist[1]
+                to_kill_first: List[Process]
+                to_kill_last: List[Process]
+                to_kill_first, to_kill_last = self.get_hitlist_in_order()
 
                 self.termination_event.set()
 
-                # to make sure we only warn the user once about
-                # the pending modules
-                self.warning_printed_once = False
-
                 try:
                     # Wait timeout_seconds for all the processes to finish
-                    while time.time() - method_start_time < timeout_seconds:
+                    while time.time() - method_start_time < timeout:
                         (
                             to_kill_first,
                             to_kill_last,
@@ -701,7 +734,7 @@ class ProcessManager:
                     reason = "User pressed ctr+c or Slips was killed by the OS"
                     graceful_shutdown = False
 
-                if time.time() - method_start_time >= timeout_seconds:
+                if time.time() - method_start_time >= timeout:
                     # getting here means we're killing them bc of the timeout
                     # not getting here means we're killing them bc of double
                     # ctr+c OR they terminated successfully
@@ -740,7 +773,7 @@ class ProcessManager:
             self.main.profilers_manager.cpu_profiler_release()
             self.main.profilers_manager.memory_profiler_release()
 
-            self.main.db.close()
+            self.main.db.close_redis_and_sqlite()
             if graceful_shutdown:
                 print(
                     "[Process Manager] Slips shutdown gracefully\n",