+++ slips_files/common/parsers/config_parser.py
@@ -722,3 +722,25 @@ class ConfigParser(object):
         return self.read_configuration(
             "global_p2p", "iris_conf", "config/iris_config.yaml"
         )
+
+    def get_bootstrapping_setting(self) -> (bool, list):
+        return (
+            self.read_configuration("global_p2p", "bootstrapping_node", False)
+            and self.read_configuration("global_p2p", "use_global_p2p", False)
+            and ("-i" in sys.argv or "-g" in sys.argv),
+            ["fidesModule", "irisModule"],
+        )
+
+    def is_bootstrapping_node(self) -> bool:
+        return (
+            self.read_configuration("global_p2p", "bootstrapping_node", False)
+            and self.read_configuration("global_p2p", "use_global_p2p", False)
+            and ("-i" in sys.argv or "-g" in sys.argv)
+        )
+
+    def get_bootstrapping_modules(self) -> list:
+        return self.read_configuration(
+            "global_p2p",
+            "bootstrapping_modules",
+            ["fidesModule", "irisModule"],
+        )