+++ slips_files/core/database/sqlite_db/database.py
@@ -48,7 +48,9 @@ class SQLiteDB:
         self.conn = sqlite3.connect(
             self._flows_db, check_same_thread=False, timeout=20
         )
-
+        # enable write-ahead logging for concurrent reads and writes to
+        # avoid the "DB is locked" error
+        self.conn.execute("PRAGMA journal_mode=WAL;")
         self.cursor = self.conn.cursor()
         if db_newly_created:
             # only init tables if the db is newly created
@@ -392,7 +394,7 @@ class SQLiteDB:
         Each call to this function results in 1 sqlite transaction
         """
         trial = 0
-        max_trials = 3
+        max_trials = 5
         while trial < max_trials:
             try:
                 # note that self.conn.in_transaction is not reliable
@@ -428,7 +430,6 @@ class SQLiteDB:
                         f"Query discarded.",
                         0,
                         1,
-                        log_to_logfiles_only=True,
                     )
                     return