+++ tests/module_factory.py
@@ -679,6 +679,9 @@ class ModuleFactory:
     def create_process_manager_obj(self):
         main_mock = Mock()
         main_mock.conf.get_disabled_modules.return_value = []
+        #main_mock.conf.get_bootstrapping_setting.return_value = (False, [])
+        main_mock.conf.is_bootstrapping_node.return_value = False
+        main_mock.conf.get_bootstrapping_modules.return_value = ["fidesModule", "irisModule"]
         main_mock.input_type = "pcap"
         main_mock.mode = "normal"
         main_mock.stdout = ""