+++ tests/test_flow_handler.py
@@ -304,10 +304,8 @@ def test_handle_dhcp():
         starttime=1234567890,
         uids=["uid1", "uid2", "uid3"],
         smac="aa:bb:cc:dd:ee:ff",
-        saddr="192.168.1.1",
-        server_addr="192.168.1.1",
-        daddr="192.168.1.2",
-        client_addr="192.168.1.3",
+        server_addr="192.168.1.2",
+        client_addr="192.168.1.1",
         host_name="test-host",
         requested_addr="192.168.1.4",
     )
@@ -319,7 +317,7 @@ def test_handle_dhcp():
     flow_handler.db.add_mac_addr_to_profile.assert_called_with(
         flow_handler.profileid, flow.smac
     )
-    flow_handler.db.store_dhcp_server.assert_called_with("192.168.1.1")
+    flow_handler.db.store_dhcp_server.assert_called_with("192.168.1.2")
     flow_handler.db.mark_profile_as_dhcp.assert_called_with(
         flow_handler.profileid
     )