+++ tests/test_slips_utils.py
@@ -16,7 +16,7 @@ def test_get_sha256_hash():
     utils = ModuleFactory().create_utils_obj()
     # a file that we know doesn't change
     assert (
-        utils.get_sha256_hash("modules/template/__init__.py")
+        utils.get_sha256_hash_of_file_contents("modules/template/__init__.py")
         == "683de4e72614dd4947e5f3b5889e12fa15bf6d5b4c5978683bad78f3c6ad5695"
     )
 
@@ -24,14 +24,45 @@ def test_get_sha256_hash():
 def test_get_sha256_hash_from_nonexistent_file():
     utils = ModuleFactory().create_utils_obj()
     with pytest.raises(FileNotFoundError):
-        utils.get_sha256_hash("nonexistent_file.txt")
+        utils.get_sha256_hash_of_file_contents("nonexistent_file.txt")
+
+
+@pytest.mark.parametrize(
+    "filepath, expected_result",
+    [  # Testcase 1: Supported file
+        ("path/to/conn.log", False),
+        ("path/to/dns.log", False),
+        ("path/to/http.log", False),
+        ("path/to/ssl.log", False),
+        ("path/to/ssh.log", False),
+        ("path/to/dhcp.log", False),
+        ("path/to/ftp.log", False),
+        ("path/to/smtp.log", False),
+        ("path/to/tunnel.log", False),
+        ("path/to/notice.log", False),
+        ("path/to/files.log", False),
+        ("path/to/arp.log", False),
+        ("path/to/software.log", False),
+        ("path/to/software.log.labeled", False),
+        ("path/to/weird.log", False),
+        ("path/to/software.log.labeled.something", True),
+        ("path/to/unsupported.log", True),
+    ],
+)
+def test_is_ignored_zeek_log_file(filepath, expected_result):
+    """
+    Test that the is_ignored_file method correctly
+    identifies ignored Zeek log files.
+    """
+    utils = ModuleFactory().create_utils_obj()
+    assert utils.is_ignored_zeek_log_file(filepath) == expected_result
 
 
 def test_get_sha256_hash_permission_error():
     utils = ModuleFactory().create_utils_obj()
     with patch("builtins.open", side_effect=PermissionError):
         with pytest.raises(PermissionError):
-            utils.get_sha256_hash("restricted_file.txt")
+            utils.get_sha256_hash_of_file_contents("restricted_file.txt")
 
 
 @pytest.mark.parametrize(