+++ tests/test_update_file_manager.py
@@ -92,7 +92,7 @@ def test_check_if_update_local_file(
     update_manager = ModuleFactory().create_update_manager_obj()
 
     mocker.patch(
-        "slips_files.common.slips_utils.Utils.get_sha256_hash",
+        "slips_files.common.slips_utils.Utils.get_sha256_hash_of_file_contents",
         return_value=new_hash,
     )
     update_manager.db.get_ti_feed_info.return_value = {"hash": old_hash}
@@ -671,7 +671,8 @@ def test_check_if_update_org(
 
     update_manager.db.get_ti_feed_info.return_value = {"hash": cached_hash}
     mocker.patch(
-        "slips_files.common." "slips_utils.Utils.get_sha256_hash",
+        "slips_files.common."
+        "slips_utils.Utils.get_sha256_hash_of_file_contents",
         return_value=hash(file_content.encode()),
     )
     result = update_manager.check_if_update_org("test_org")